+++ test/state_set_test.dart
@@ -1,5 +1,0 @@
-import 'package:flutter_test/flutter_test.dart';
-
-import 'package:state_set/state_set.dart';
-
-void main() {}